Start Replacing Overworlds
Started putting in Crystal's overworld sprites. Palettes are in and working. Need to figure out how to load the palettes based on a byte in the event object structure and not on the default palette assigned to it. Also need to see how Frontier Trainers are linked to their Overworlds.