Override OW Palettes and More OWs Added
The palette slot an OW uses can now be manually overwritten to change its color. The unk2 field of the EventObjectTemplate struct is used to indicate this. If it is zero (the default), it will use the OW's normal palette. If it is anything else, it will use that palette slot, starting with the slot 2 (0 is the player and 1 is its reflection). So, putting 1 in unk2 will get the third palette slot, 2 will get the fourth, etc. In a map's events.inc, this is the value right after its EVENT_OBJ_GFX field.

In addition, replaced more OW sprites and made some edits to maps to ensure they and their color changes are working correctly.