Add Remaining OW Sprites
All OW sprites from Crystal are in, save for the sprites used in the party Pokemon screen that can show up in the OW. Leaf has her own sprite for linking, but no flavors of Brendan or May do. In addition, OWs have been named appropriately and OWs unused in GSC have been moved to the bottom of the alphabetical list that porymap shows for ease of use.

Care was taken to replace OWs for like Trainer classes with like Trainer classes where possible, in case the Frontier is reused. ExpertF is now Teacher. ExpertM is now Sage. HexManiac is now Medium.