Add All Trainer Pics
All enemy Trainer front pics are in. Some facility trainers have been reclassed to accomodate some of the overwrites needed. All Swimming Triathletes -> Swimmers, for example. The end result is that Johto classes that aren't direct equivalents will not mess with the Frontier (except ExpertM -> Sage, ExpertF -> Teacher, Hex Maniac -> Medium).

Still need front pics for Leaf, and both flavors of Brendan and May. Not counting the Frontier Brains (which may end up used if I decide on keeping the Frontier as an option) and the aforementioned player characters, there are 15 regular Trainers that still have GBA Trainer sprites. They will need to be reclassed in the Frontier as well.