Rename Maps up to Route 34 & Map MooMoo Farm Barn
Routes 33 and 34 have been renamed, as well as Azalea Town and all of its houses. MooMoo Farm's Barn is mapped, finishing out all indoor maps on Johto routes that aren't dungeons.

From this point on, none of the towns' houses are mapped or set up, so renaming will be less tedious. Will only be renaming the towns themselves, their Pokemon Centers, and their Marts. The remaining maps will just be new ones to hopefully avoid the drudgery of renaming.