Map and Event All of Ecruteak City
All of indoor and outdoor Ecruteak is evented, and indoor Ecruteak maps have been added. This does not include the Burned and Tin Towers, as they are dungeons. The groups have been rearranged again to accommodate the Ecruteak maps.

Additionally, Teleport can now be used anywhere for debugging purposes. TODO.txt has been updated with a reminder to remove that for release.