Event and map Routes 44, 45, 46, and Blackthorn City
Routes 44, 45, 46, and all of indoor and outdoor Blackthorn City is evented. Indoor Blackthorn is mapped save for the Dragon's Den, which is a dungeon.

Evented the overlooked Violet and Azalea Gyms.

All of Johto is now mapped and evented outside of dungeons, the Battle Tower areas, Olivine Port, and New Bark Town (have to mess with starting scripts first).