Rename & Event National Park & Tin Tower, Map & Event 2/3rds of Dragons Den
National Park and Tin Tower now have proper names and are evented. Tin Tower's Roof has been moved to the outdoor tileset because of the time of day transitions.

Dragon's Den B1F and the Dragon Shrine is mapped, evented, and properly named. The cave entrance to the Dragon's Den is not yet mapped, as the cave tileset is not in yet.

The dungeon map group is not yet sorted properly.