Add Indoor Port Tileset; Map Burned Tower B1F
The Cerulean Gym portion of the Port tileset has been added to the end of the GenericBuilding tileset. The Building and GenericBuilding tilesets had some tiles shuffled around to accommodate for them. The remaining portions of the Port tileset has to go on the outdoor tileset due to the Olivine and Vermillion Port maps cycling palettes by time of day.

Burned Tower B1F is mapped.

Updated TODO.txt again.