Add Ruins Of Alph Tileset; Map & Event Ruins Of Alph
A new primary tileset (Special) has been added and will contain tilesets with custom palettes (Ice Path, Radio Tower, etc.). As of now, only Ruins of Alph (which uses dungeon day palette and couldn't fit in other tilesets) is blocked out on it.

All Ruins of Alph chambers as well as the research center are mapped and evented. The message rooms probably need to be translated for each language.