Port & Battle Tower Exterior Tilesets Added; Map Olivine Port & Battle Tower Outside
The outdoor portion of the Port tileset, as well as the outdoor Battle Tower tileset have been added to the outdoor tileset. There are no more tilesets to import.

Olivine Port has been mapped. It was already evented.

The Battle Tower's Exterior is mapped and evented.

Fixed an incorrect roof on Route 40.

Updated TODO.txt.