Map & Event Remaining Johto and Pre-Fast Ship Areas
Routes 23, 26, 27, and 28 are evented. The houses on these Routes are mapped and evented.

Tohjo Falls, Victory Road Gate, Victory Road, Indigo Plateau Pokemon Center, the Elite Four Rooms, and Lance's Room are mapped and evented.

The Hall of Fame room is evented but not mapped, as its tileset is not ready.

With the exception of the Hall of Fame, all maps up to the arrival in Vermillion City are finished.