Start Trainer Case
The first page of the Trainer Case is working correctly. Front pics for the Crystal male and female player still need to be added; Gold's player sprite is used as a placeholder.

The GSC Trainer Case was added as a separate type of Trainer Card, so the RSEFRLG Trainer Cards have been preserved for linking purposes. The only change that would affect them is the location of the Trainer sprite, as that layer's position was modified. They would still need to be devamped, regardless.

Additionally, added some new strings to strings.c for the Trainer Case, another colon to the font and charmap, and changed the menu option for the PokeNav to PokeGear.