Import Gold/Silver Level-Up Learnsets
The Gold and Silver level-up learnsets are in as a separate file. No functionality to swap between them has been implemented.

It seems gLevelUpLearnsets holds the learnset data that the move-learning-related functions use. It is defined in src/data/level_up_learnset_pointers.h. Will need to make it a 2D array or just make a new array for the GS ones and have if statements to swap between them.