Deletion Spree and Initial Flag Work
Removed most unused tilesets, maps, and layouts. Things that may see use in the future were preserved. Also disabled ambient cries on maps with Pokemon because I found that while I was in there.

The version flags (FLAG_IS_VERSION_GOLD and so forth) are defined, but not yet used.