Start Pokegear
Pokegear pages are in and (barely) functional as a menu. The Time Card is fully functional as of now, though without a way to set the day of the week, it will display garbage text. There is no functionality for the other cards. Pages scroll according to which cards you have, and the Map Card will switch between displaying Johto and Kanto as it should depending on the area.

Additionally, did more clean up from the deletion spree.

BUG: The text on the Time Card sometimes flickers. Check flags to prevent rewriting it?
BUG: The day of the week text may leave behind artifacts when it is updated.
BUG: There is something weird happening with how the day of the week changes. To be investigated.